Refactor checkForEnd and updating comments
Refactoring checkForEnd to store full game logic in cage and custom level end logic
